% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aggregation_table.R
\name{build_aggregation_table_validation}
\alias{build_aggregation_table_validation}
\title{Validation: Build aggregation table}
\usage{
build_aggregation_table_validation(
  polygons,
  id_raster,
  polygon_id_field,
  polygon_ids
)
}
\arguments{
\item{polygons}{terra SpatVector object. Should contain a unique ID field.}

\item{id_raster}{terra SpatRaster object. ID raster created by \code{build_id_raster()} for
the polygons object. Should have the same CRS as \code{polygons} and completely cover it.}

\item{polygon_id_field}{(character) Unique identifier field in \code{polygons}.}

\item{polygon_ids}{(vector, any type) Polygon identifiers from \code{polygon_id_field}.}
}
\value{
Errors if checks fail; silently passes if checks pass
}
\description{
Input data validation for \code{build_aggregation_table()}
}
\seealso{
build_aggregation_table
}
