% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_world_template_raster.R
\name{make_world_template_raster}
\alias{make_world_template_raster}
\title{Make world template raster}
\usage{
make_world_template_raster()
}
\value{
terra SpatRaster matching the details above
}
\description{
Create a template raster for the world with approximately 5x5km resolution
at the equator, matching many common raster covariates for health.
}
\details{
The raster has the following specifications:
\itemize{
\item 4320 rows, 8640 columns
\item Resolution: 0.04166667 decimal degrees, approx. 5km at the equator
\item CRS: WGS 1984 unprojected latitude/longitude, \code{terra::crs('EPSG:4326')}
\item Values: All NA. Used exclusively for creating a shapefile-specific ID raster
}
}
