% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aggregation_table.R
\name{build_aggregation_table}
\alias{build_aggregation_table}
\title{Build aggregation table}
\usage{
build_aggregation_table(polygons, id_raster, polygon_id_field, verbose = TRUE)
}
\arguments{
\item{polygons}{terra SpatVector object. Should contain a unique ID field.}

\item{id_raster}{terra SpatRaster object. ID raster created by \code{build_id_raster()} for
the polygons object. Should have the same CRS as \code{polygons} and completely cover it.}

\item{polygon_id_field}{(character) Unique identifier field in \code{polygons}.}

\item{verbose}{(logical) Show progress for building aggregation rows for each polygon?}
}
\value{
data.table with fields:
\itemize{
\item polygon_id: Unique polygon identifier
\item pixel_id: unique pixel ID from the ID raster
\item masked_pixel_id: Index counting only non-NA pixels from the ID raster
\item area_fraction: fraction of the pixel area falling within this polygon
\item Merged fields from the table of polygons
}
}
\description{
Build a table to quickly aggregate from pixels to polygons
}
\seealso{
calculate_pixel_fractions_single_polygon()
}
